Fix OOB/UAF/leaks and reentrancy issues in PCSX2 core
CDVD: bounds check mg_BIToffset loop to prevent OOB read on
crafted MG data (high severity)
ThreadedFileReader: use temp pointer for realloc to prevent
leak on allocation failure
main.cpp: replace strncpy with strlcpy for guaranteed null-
termination
Vif_Codes: use u64 arithmetic in MPG size check to prevent
overflow bypass
R5900: make eeEventTestIsActive atomic for thread-safe
interrupt path access
Console: use recursive_mutex to prevent deadlock on
reentrant Console.Error calls

+++ pcsx2/Vif_Codes.cpp
@@ -303,7 +303,7 @@ static __fi void _vifCode_MPG(int idx, u32 addr, const u32* data, int size)
 
 	if (idx && THREAD_VU1)
 	{
-		if ((addr + size * 4) > vuMemSize)
+		if ((static_cast<u64>(addr) + static_cast<u64>(size) * 4) > vuMemSize)
 		{
 			vu1Thread.WriteMicroMem(addr, (u8*)data, vuMemSize - addr);
 			size -= (vuMemSize - addr) / 4;
@@ -320,7 +320,7 @@ static __fi void _vifCode_MPG(int idx, u32 addr, const u32* data, int size)
 	}
 
 	// Don't forget the Unsigned designator for these checks
-	if ((addr + size * 4) > vuMemSize)
+	if ((static_cast<u64>(addr) + static_cast<u64>(size) * 4) > vuMemSize)
 	{
 		if (!idx)
 			CpuVU0->Clear(addr, vuMemSize - addr);
